Lab 1 done
Wykonanie pierwszych labów, liczenie średniej ocen dla semestrów, wypisanie przedmiotu z najwyższą i najniższą oceną, zmiana zmiennych z int na float

+++ lab1/src/CustomClasses/Zapis_c.java
@@ -9,7 +9,7 @@ import java.util.stream.Stream;
 
 public class Zapis_c {
 
-    public static void ZStreamu(String filePath, Stream<Map.Entry<String,Integer>> arr) {
+    public static void ZStreamu(String filePath, Stream<Map.Entry<String,Float>> arr) {
         File file = new File(filePath);
         try{
             FileWriter fileWriter = new FileWriter(file,true);
